requirements-full.txt
# Production requirements for Python <= 3.12
# Use this file if you want full background removal functionality

streamlit>=1.28.0
Pillow>=10.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
scikit-image>=0.21.0
scipy>=1.10.0
